+++ secure/js/serverlist.js
@@ -69,7 +69,7 @@ const _clearError = () => {
   while (errorDivEl.firstChild) {
     errorDivEl.removeChild(errorDivEl.firstChild);
   }
-  document.getElementById('showRefreshButtonDiv').setAttribute('hidden', '');
+  document.getElementById('showForceButtonDiv').setAttribute('hidden', '');
 };
 
 /**
@@ -82,7 +82,7 @@ const _showError = (errorString) => {
   const errorMessageEl = document.createElement('div');
   errorMessageEl.textContent = errorString || 'Error: unknown error (4712)';
   errorDivEl.appendChild(errorMessageEl);
-  document.getElementById('showRefreshButtonDiv').removeAttribute('hidden');
+  document.getElementById('showForceButtonDiv').removeAttribute('hidden');
   // scroll to bottom of page to show error
   window.scrollTo(0, document.querySelector('body').scrollHeight);
 };
@@ -584,11 +584,16 @@ const buildServerListTable = (data) => {
     // Dynamically create table elements and add to table
     //
     if ((Array.isArray(data)) && (data.length > 0)) {
+      document.getElementById('tableVisibilityDiv').removeAttribute('hidden');
+      document.getElementById('emptyTableDiv').setAttribute('hidden', '');
+      let allServersDisabled = true;
       for (let i = 0; i < data.length; i++) {
         const rowEl = document.createElement('tr');
         rowEl.setAttribute('index', i.toString());
         if (data[i].disabled) {
           rowEl.classList.add('disabled-tr');
+        } else {
+          allServersDisabled = false;
         }
 
         const td01El = document.createElement('td');
@@ -824,6 +829,22 @@ const buildServerListTable = (data) => {
         } // if editable
         tableNode.appendChild(rowEl);
       } // next i
+      if (allServersDisabled) {
+        document.getElementById('allDisabledWarningDiv').removeAttribute('hidden');
+        document.getElementById('returnIrcButton').setAttribute('disabled', '');
+        document.getElementById('fullButton').removeAttribute('disabled');
+      } else {
+        document.getElementById('allDisabledWarningDiv').setAttribute('hidden', '');
+        document.getElementById('returnIrcButton').removeAttribute('disabled');
+        document.getElementById('fullButton').removeAttribute('disabled');
+      }
+    } else {
+      // else case of empty server list, show instructions
+      document.getElementById('tableVisibilityDiv').setAttribute('hidden', '');
+      document.getElementById('emptyTableDiv').removeAttribute('hidden');
+      document.getElementById('allDisabledWarningDiv').setAttribute('hidden', '');
+      document.getElementById('returnIrcButton').setAttribute('disabled', '');
+      document.getElementById('fullButton').setAttribute('disabled', '');
     }
     resolve(null);
   });
@@ -851,24 +872,25 @@ const checkForApiError = (data) => {
  * @returns {Promise} Resolved to null
  */
 const setDivVisibility = (data) => {
+  // console.log('setDivVisibility data: ', JSON.stringify(data, null, 2));
   document.getElementById('listVisibilityDiv').removeAttribute('hidden', '');
   document.getElementById('formVisibilityDiv').setAttribute('hidden', '');
   document.getElementById('serverPasswordWarningDiv').setAttribute('hidden', '');
   document.getElementById('nickservCommandWarningDiv').setAttribute('hidden', '');
   // show/hide buttons in server list table
   if ((data.ircConnected) || (data.ircConnecting)) {
-    document.getElementById('createNewButton').setAttribute('hidden', '');
+    document.getElementById('createNewButton').setAttribute('disabled', '');
     document.getElementById('warningVisibilityDiv').removeAttribute('hidden');
     editable = false;
   } else {
-    document.getElementById('createNewButton').removeAttribute('hidden');
+    document.getElementById('createNewButton').removeAttribute('disabled');
     document.getElementById('warningVisibilityDiv').setAttribute('hidden', '');
     editable = true;
   }
   // show/hide proxy setting info at bottom
   if (data.enableSocks5Proxy) {
     document.getElementById('ircProxyDiv').textContent =
-      'Socks5 Proxy: Enabled Globally\nSocks5 Proxy: ' +
+      'Socks5 Proxy: Enabled Globally  Socks5 Proxy: ' +
       data.socks5Host + ':' + data.socks5Port;
   } else {
     document.getElementById('ircProxyDiv').textContent = 'Socks5 Proxy: Disabled Globally';
@@ -881,13 +903,188 @@ const setDivVisibility = (data) => {
 // -------------------------------------
 
 /**
- * Show help for server groups
+ * Show help for server inputs
  */
-document.getElementById('groupInfoButton').addEventListener('click', () => {
-  if (document.getElementById('groupInfoHiddenDiv').hasAttribute('hidden')) {
-    document.getElementById('groupInfoHiddenDiv').removeAttribute('hidden');
+document.getElementById('disabledCheckboxInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('disabledCheckboxInfo').hasAttribute('hidden')) {
+    document.getElementById('disabledCheckboxInfo').removeAttribute('hidden');
   } else {
-    document.getElementById('groupInfoHiddenDiv').setAttribute('hidden', '');
+    document.getElementById('disabledCheckboxInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('groupInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('groupInputInfo').hasAttribute('hidden')) {
+    document.getElementById('groupInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('groupInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('nameInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('nameInputInfo').hasAttribute('hidden')) {
+    document.getElementById('nameInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('nameInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('autoReconnectCheckboxInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('autoReconnectCheckboxInfo').hasAttribute('hidden')) {
+    document.getElementById('autoReconnectCheckboxInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('autoReconnectCheckboxInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('loggingCheckboxInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('loggingCheckboxInfo').hasAttribute('hidden')) {
+    document.getElementById('loggingCheckboxInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('loggingCheckboxInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('hostInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('hostInputInfo').hasAttribute('hidden')) {
+    document.getElementById('hostInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('hostInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('proxyCheckboxInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('proxyCheckboxInfo').hasAttribute('hidden')) {
+    document.getElementById('proxyCheckboxInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('proxyCheckboxInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('ircProxyInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('ircProxyInfo').hasAttribute('hidden')) {
+    document.getElementById('ircProxyInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('ircProxyInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('passwordInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('passwordInputInfo').hasAttribute('hidden')) {
+    document.getElementById('passwordInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('passwordInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('saslUsernameInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('saslUsernameInputInfo').hasAttribute('hidden')) {
+    document.getElementById('saslUsernameInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('saslUsernameInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('saslPasswordInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('saslPasswordInputInfo').hasAttribute('hidden')) {
+    document.getElementById('saslPasswordInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('saslPasswordInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('nickInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('nickInputInfo').hasAttribute('hidden')) {
+    document.getElementById('nickInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('nickInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('altNickInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('altNickInputInfo').hasAttribute('hidden')) {
+    document.getElementById('altNickInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('altNickInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('recoverNickCheckboxInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('recoverNickCheckboxInfo').hasAttribute('hidden')) {
+    document.getElementById('recoverNickCheckboxInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('recoverNickCheckboxInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('realInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('realInputInfo').hasAttribute('hidden')) {
+    document.getElementById('realInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('realInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('userInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('userInputInfo').hasAttribute('hidden')) {
+    document.getElementById('userInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('userInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('identifyNickInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('identifyNickInputInfo').hasAttribute('hidden')) {
+    document.getElementById('identifyNickInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('identifyNickInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('modesInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('modesInputInfo').hasAttribute('hidden')) {
+    document.getElementById('modesInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('modesInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('channelListInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('channelListInputInfo').hasAttribute('hidden')) {
+    document.getElementById('channelListInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('channelListInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('identifyCommandInputInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('identifyCommandInputInfo').hasAttribute('hidden')) {
+    document.getElementById('identifyCommandInputInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('identifyCommandInputInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('forceUnlockAllInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('forceUnlockAllInfo').hasAttribute('hidden')) {
+    document.getElementById('forceUnlockAllInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('forceUnlockAllInfo').setAttribute('hidden', '');
+  }
+});
+
+document.getElementById('logoutLinkInfoBtn').addEventListener('click', () => {
+  if (document.getElementById('logoutLinkInfo').hasAttribute('hidden')) {
+    document.getElementById('logoutLinkInfo').removeAttribute('hidden');
+  } else {
+    document.getElementById('logoutLinkInfo').setAttribute('hidden', '');
+  }
+});
+document.getElementById('logoutLinkInfoBtn2').addEventListener('click', () => {
+  if (document.getElementById('logoutLinkInfo2').hasAttribute('hidden')) {
+    document.getElementById('logoutLinkInfo2').removeAttribute('hidden');
+  } else {
+    document.getElementById('logoutLinkInfo2').setAttribute('hidden', '');
   }
 });
 
@@ -1073,3 +1270,31 @@ fetchIrcState()
       console.log(err);
     }
   });
+
+//
+// The server can be optionally configured to show the /docs folder
+// If /docs is enabled, API active, then make button visible
+//
+document.getElementById('showHelpDocsButton').setAttribute('hidden', '');
+const docsTestUrl = '/irc/docs/index.html';
+const fetchOptions = {
+  method: 'HEAD',
+  headers: {
+    Accept: 'text/html'
+  }
+};
+fetch(docsTestUrl, fetchOptions)
+  .then((response) => {
+    // console.log(response.status);
+    if (response.ok) {
+      // Make the irc-hybrid-client documentation link button visible
+      document.getElementById('showHelpDocsButton').removeAttribute('hidden');
+    } else {
+      document.getElementById('showHelpDocsButton').setAttribute('hidden', '');
+    }
+  });
+
+// Hide the notes div after timeout
+setTimeout(function () {
+  document.getElementById('hidableStartupNotes').setAttribute('hidden', '');
+}, 5000);